+++ src/Traits/CanAuthorizeWithCustomer.php
@@ -26,6 +26,7 @@ namespace heidelpayPHP\Traits;
 
 use heidelpayPHP\Exceptions\HeidelpayApiException;
 use heidelpayPHP\Interfaces\HeidelpayParentInterface;
+use heidelpayPHP\Resources\Basket;
 use heidelpayPHP\Resources\Customer;
 use heidelpayPHP\Resources\Metadata;
 use heidelpayPHP\Resources\TransactionTypes\Authorization;
@@ -42,6 +43,9 @@ trait CanAuthorizeWithCustomer
      * @param Customer|string $customer
      * @param string|null     $orderId
      * @param Metadata|null   $metadata
+     * @param Basket|null     $basket   The Basket object corresponding to the payment.
+     *                                  The Basket object will be created automatically if it does not exist
+     *                                  yet (i.e. has no id).
      *
      * @return Authorization
      *
@@ -54,7 +58,8 @@ trait CanAuthorizeWithCustomer
         $returnUrl,
         $customer,
         $orderId = null,
-        $metadata = null
+        $metadata = null,
+        $basket = null
     ): Authorization {
         if ($this instanceof HeidelpayParentInterface) {
             return $this->getHeidelpayObject()->authorize(
@@ -64,7 +69,8 @@ trait CanAuthorizeWithCustomer
                 $returnUrl,
                 $customer,
                 $orderId,
-                $metadata
+                $metadata,
+                $basket
             );
         }