+++ CHANGELOG.md
@@ -3,6 +3,21 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.1.0][1.0.1.0]
+
+### Added
+*   It is now possible to create, update and fetch a basket as well as referencing it by a authorization or charge.
+*   Extended facade to enable basket handling.
+*   Missing tests for metadata resource.
+
+### Changed
+*   Refactor value update to allow for empty strings.
+*   Ensuring that transferred floats are always encoded as floats on json-encode.
+*   Properties stored in an array are now (json-)encoded as \stdClass not as array.
+
+### Fixed
+*   Comments and styles.
+
 ## [1.0.0.1][1.0.0.1]
 
 ### Fixed
@@ -73,3 +88,4 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 [1.0.0-beta.2]: https://github.com/heidelpay/heidelpayPHP/compare/1.0.0-beta.1..1.0.0-beta.2
 [1.0.0.0]: https://github.com/heidelpay/heidelpayPHP/compare/1.0.0-beta.2..1.0.0.0
 [1.0.0.1]: https://github.com/heidelpay/heidelpayPHP/compare/1.0.0.0..1.0.0.1
+[1.0.1.0]: https://github.com/heidelpay/heidelpayPHP/compare/1.0.0.1..1.0.1.0